moved chunking and cleaning text from pdf_parser to ingest_utils
