+++ ingest/ingester.py
@@ -8,6 +8,7 @@ from loguru import logger
 import settings
 from .pdf_parser import PdfParser
 from .content_iterator import ContentIterator
+from ingest.ingest_utils import IngestUtils
 
 
 class Ingester:
@@ -31,19 +32,25 @@ class Ingester:
         content_iterator = ContentIterator(self.content_folder)
         # create text chunks with chosen settings of chunk size and chunk overlap
         pdf_parser = PdfParser(self.chunk_size, self.chunk_overlap, self.file_no)
+        ingestutils = IngestUtils(self.chunk_size, self.chunk_overlap, self.file_no)
+        # txt_parser = 
 
         chunks: List[docstore.Document] = []
         # for each pdf file that the content_iterator yields
         for document in content_iterator:
             if document.endswith(".pdf"):
-                # convert the pdf text to cleaned text chunks
+                # check document path
                 pdf_parser.set_pdf_file_path(document)
-                document_chunks = pdf_parser.clean_text_to_docs()
-                chunks.extend(document_chunks)
-                logger.info(f"Extracted {len(chunks)} chunks from {document}")
+                raw_pages, metadata = pdf_parser.parse_pdf()
+
             else:
                 logger.info(f"Cannot ingest document {document} because it has extension {document[-4:]}")
-        
+
+            # convert the raw text to cleaned text chunks
+            document_chunks = ingestutils.clean_text_to_docs(raw_pages, metadata)
+            chunks.extend(document_chunks)
+            logger.info(f"Extracted {len(chunks)} chunks from {document}")
+
         if self.embeddings_provider == "openai":
             embeddings = OpenAIEmbeddings(model=self.embeddings_model, client=None)
             logger.info("Loaded openai embeddings")
